pyproject.toml
[project]
name = "rpifan"
version = "0.2.0"
description = "Raspberry Pi Smart Fan Controller with TUI (PID + PWM)"
readme = "README.md"
authors = [{name = "Snow", email = "snow@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"

dependencies = [
    "RPi.GPIO>=0.7.1",
    "psutil>=7.2.2",
    "rich>=13.0.0",
]

[project.scripts]
fan-backend = "src.backend:main"
fan-ui = "src.frontend:main"
